Update nullable bool to be single widget
